+++ configs/react-hooks.js
@@ -1,6 +1,0 @@
-// eslint-plugin-react-hooks doesn't export a recommended config
-module.exports = {
-  rules: {
-    'healthier/rules-of-hooks': 'error'
-  }
-}