+++ configs/standard-jsx.js
@@ -1,1 +1,0 @@
-module.exports = require('./convert')(require('eslint-config-standard-jsx'))